Rename prefillcards package to testcards.
