+++ Demo/src/main/java/com/paypal/android/ui/testcards/PrefillCardsFragment.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.prefillcards
+package com.paypal.android.ui.testcards
 
 import android.os.Bundle
 import android.view.LayoutInflater
