+++ Demo/src/main/res/layout/fragment_prefill_cards.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.prefillcards.PrefillCardsFragment">
+    tools:context=".ui.testcards.PrefillCardsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
