+++ Demo/src/main/res/navigation/nav_graph.xml
@@ -45,7 +45,7 @@
         android:label="PayPalButtonsFragment" />
     <fragment
         android:id="@+id/prefillCardFragment"
-        android:name="com.paypal.android.ui.prefillcards.PrefillCardsFragment"
+        android:name="com.paypal.android.ui.testcards.PrefillCardsFragment"
         android:label="Prefill Card"
         tools:layout="@layout/fragment_prefill_cards" />
 </navigation>
